Move logic from Tetronimo subclasses to superclass
In the previous commmits, essentially the same code
was used in all of the Tetronimo subclasses with few variations.
This lead both to uneccessary code and defeated the point of
extending from an abstract class in the first place. This
change should make the code less messy and easier to maintain.